+++ src/importer/importer_optimized.py
@@ -76,6 +76,7 @@ from src.importer.importer import (
     validate_jg_record,
     validate_se_record,
     validate_sk_record,
+    validate_h1_record,
     validate_um_record,
     validate_tc_record,
     validate_wc_record,
@@ -101,6 +102,7 @@ from src.importer.importer import (
     verify_sk_storage_schema,
     verify_tc_storage_schema,
     verify_tk_coupled_tables,
+    verify_h1_storage_schema,
     verify_um_storage_schema,
     verify_wc_storage_schema,
     verify_we_storage_schema,
@@ -156,6 +158,7 @@ class OptimizedDataImporter:
         self._verified_jc_tables: set[str] = set()
         self._verified_cs_tables: set[str] = set()
         self._verified_um_tables: set[str] = set()
+        self._verified_h1_tables: set[str] = set()
         self._verified_jg_tables: set[str] = set()
         self._verified_wc_tables: set[str] = set()
         self._verified_wf_tables: set[str] = set()
@@ -498,6 +501,10 @@ class OptimizedDataImporter:
                     if verify_um_storage_schema(self.database, table_name):
                         self._verified_um_tables.add(table_name)
                 validate_um_record(record, table_name)
+                if table_name not in self._verified_h1_tables:
+                    if verify_h1_storage_schema(self.database, table_name):
+                        self._verified_h1_tables.add(table_name)
+                validate_h1_record(record, table_name)
                 if table_name not in self._verified_jg_tables:
                     if verify_jg_storage_schema(self.database, table_name):
                         self._verified_jg_tables.add(table_name)